+++ NEWS.md
@@ -4,6 +4,7 @@ robotoolbox 1.0.5.9000
 ### NEW FEATURES
 
 - Update to make it compatible with `tidyr` `1.3.0`
+- Hide token when using `kobo_settings` print method
 
 robotoolbox 1.0.4.9000
 ======================